requirements.txt
Django==1.8.3
Pillow==3.2.0
pytz==2016.4
djangorestframework==3.3.3
Pygments==2.1.3
